Updated hand model to add NEW HAND RIG!!!!!
The rig is fully functional with it's own skeleton, BP and AnimBP. It is still a little jank. Sorry for that, as I'm still tweaking timings and so forth. The best part is due to the modular setup I can add or tweak animations at any time and add them back in!