Week 9 level updates (#16)
* Pause menu bug fix

Pause menu can be accessed with ESC (in builds and in Standalone Game playtest mode only, because the editor overrides ESC when playing in viewport/Editor Window). Probably shouldn't go in this branch but I forgot to fix this bug, so have it here instead

* Built the lighting for reference

I think we still have some things to deal with as the overlapping lights causes artefacting on each of the shadows. I have adjusted the Post Process Material, we can mess with the Exposure and brightness here and maybe get an even lighting source?

* Started adding more props in (+Redirector fix)

Props added, redirectors fixed, that's why there's some weird changes

I can revert this if this breaks things, but UE4 shouldn't break this?

* map fulled out with what was there

i also added posters :)

* Map 2.0 rough draft implemeneted

* level rework

redone walls, ceiling and asset placements

* Texture Scaling

scaled down some of the floors and stopped the tiles from clipping through the walls, as well as some minor clipping in the level

* Beta Lighting included

God is alive and full of spite

* Art test map with LIGHTING PASS

God is dead and we killed him

* Ceiling, Bathroom and Chair Assets added

* Added Auryn's PP in

And made compatibility fixes for merge to master

* Merge fix 2

Co-authored-by: alicemcg1603 <74119389+alicemcg1603@users.noreply.github.com>
Co-authored-by: ThermosCat <98907727+ThermosCat@users.noreply.github.com>
Co-authored-by: Brendan <brendanyim31@gmail.com>
Co-authored-by: Auryn <54985160+Nyrua@users.noreply.github.com>